Hydra Role Management Updates
Fixed bugs and added features.  Default Roles have been added for Department and Faculty (no_department, no_faculty respectively), this ensures that if a Person has a faculty code or department ou that doesn't match with any
within the Roles table, then they will default to no_* role for department or faculty.

Added Sample code into seeds.rb to guide how to create sample 'People', and Roles records for development purposes.

Test coverage added for Person/User/Role/RoleType models.

+++ db/seeds.rb
@@ -1,39 +1,51 @@
 # This file should contain all the record creation needed to seed the database with its default values.
 # The data can then be loaded with the rake db:seed (or created alongside the db with db:setup).
-#
-# Examples:
-#
-#   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
-#   Mayor.create(name: 'Emanuel', city: cities.first)
 
 #hyhull Roles
+#Clear existing roles in dev/test mode only.. 
+unless Rails.env.production?
+  Role.delete_all 
+  RoleType.delete_all 
+  Person.delete_all
+end
 
-#Clear existing roles
-Role.delete_all
-RoleType.delete_all
-Person.delete_all unless Rails.env.production?
-
-# Set role types..
+# Set role types - These are required for the Hyhull Application
 # hyhull: hyhull application specific role types
 # user: user role types generally 'staff/student/guest'
 # department_ou/faculty_code
 ["hyhull", "user", "department_ou", "faculty_code"].each {|rt| RoleType.create(name: rt)}
 
-
-# Seed the hyhull roles 
+# Seed the hyhull roles
+# These are the required hyhull specific roles
 [{ name:"contentAccessTeam", description: "Content access team role"}, 
   { name:"contentCreator", description: "Content creator role"},
   { name: "admin", description: "Admin role" }
 ].each{ |r| Role.create(name: r[:name], description: r[:description], role_type: RoleType.find_or_initialize_by_name("hyhull")) }
 
 # Seed the user roles
+# The standard list of user roles
 [{ name:"staff", description: "University staff role"},
  { name:"student", description: "University student role"},  
   { name:"guest", description: "University guest role"}
 ].each{ |r| Role.create(name: r[:name], description: r[:description], role_type: RoleType.find_or_initialize_by_name("user")) }
 
+# Seed the department roles
+# We need to think about seeding the full list for production..
+[{ name:"no_department", description: "No Department OU"} 
+].each{ |r| Role.create(name: r[:name], description: r[:description], role_type: RoleType.find_or_initialize_by_name("department_ou")) }
 
-# Add a test records to people
+# Seed the faculty roles
+# We need to think about seeding the full list for production..
+[{ name:"no_faculty", description: "No Faculty code"} 
+].each{ |r| Role.create(name: r[:name], description: r[:description], role_type: RoleType.find_or_initialize_by_name("faculty_code")) }
+  
+
+# **************************** #
+#  IMPORTANT - For Test #
+# **************************** #
+#                              #
+# These get loaded in the database, and important for running the Tests
+# Also loaded in development mode for extra test roles/users. 
 unless Rails.env.production?
   # For test lets seed some made up departments and faculities (these need to exist in the Roles DB to be picked up on user login)
 
@@ -71,5 +83,44 @@ unless Rails.env.production?
          person.faculty_code = p[:faculty_code]
          person.save
   end
+
+end
+
+# **************************** #
+#  IMPORTANT - For Development #
+# **************************** #
+#                              #
+# We will probably want to add entries to People and Roles(Departments/Faculties) that work with our development instance of CAS, i.e. users that can authenticate 
+# I.e if we authenticate with the user 'test_admin_1' who has a department of test and faculty code of '001' we will add...
+# ... a Person to the people table with a username of 'test_admin_1', and faculty_code of '001' and department_ou of 'test_dep'
+# We would also need to add faculty and department to the Roles table for them to be recognised on login. See commented example below...
+if Rails.env.development?
+
+ # ******************************************************************************************
+ # NOTE: The following is example of a Person you can add to People for development purposes
+
+ # [{ username: 'test_admin_1',  given_name: 'Test', family_name: 'Admin', email_address: 'test_admin_1@example.com', 
+ #     user_type: 'staff', department_ou: 'test_dep', faculty_code: '001'}].each do |p|
+ #         person = Person.new
+ #         person.username = p[:username]
+ #         person.given_name = p[:given_name]
+ #         person.family_name = p[:family_name]
+ #         person.email_address = p[:email_address]
+ #         person.user_type = p[:user_type]
+ #         person.department_ou = p[:department_ou]
+ #         person.faculty_code = p[:faculty_code]
+ #         person.save
+ #  end
+
+ #  # NOTE: The following is example of the corresponding Department roles you will need to add to match the above Person
+ #  [{ name:"test_dep", description: "Test Department"}].each do |r| 
+ #    Role.create(name: r[:name], description: r[:description], role_type: RoleType.find_or_initialize_by_name("department_ou")) 
+ #  end
+
+
+ #  # NOTE: The following is example of the corresponding Faculty roles you will need to add to match the above Person
+ #  [{ name:"001", description: "Test Faculty"}].each do |r|
+ #    Role.create(name: r[:name], description: r[:description], role_type: RoleType.find_or_initialize_by_name("faculty_code"))
+ #  end
 
 end
